Use only a single Person type
This is a very large refactoring change. This change was needed since
A single person, with a single ID, can be a director, writer, and actor.

All Director/Writer/Actor data types and tables were removed. Likewise, the
Directed/Wrote/Acted tables were removed as well.
A single Persons table was created. A single participations table was created,
which holds as a field the participation type.